move structures to separate folder
