+++ src/aborted/controller_spawns.js
