+++ src/controller/structures/building_factory.js
