+++ src/controller/structures/building_lab.js
