+++ src/controller/structures/building_link.js
@@ -1,4 +1,4 @@
-import { RM as _RM } from "./util_consts"
+import { RM as _RM } from "../../util_consts"
 
 /**
  *
