+++ src/controller/structures/building_observer.js
