+++ src/controller/structures/building_terminal.js
